Code cleanup
 - remove unused files
 - remove a bunch of debbuging console.logs and postMessage
 - remove unused imports
