+++ app/ts/batarangle/backend/batarangle-backend.ts
@@ -1,14 +1,3 @@
-// import {ComponentTree} from './core/mock-api';
-// import {FrontendMessagingService} from './channel/frontend-messaging-service';
-
-
-// let fms = new FrontendMessagingService();
-// window.addEventListener('message', fms.sendMessage);
-// fms.sendMessage({
-//   componentData: ct.componentTree.first()
-// });
-
-// console.log('Batarangle Backend Started!');
 chrome.runtime.onMessage.addListener((message, sender, sendResponse) => {
   console.log('Backend: Message Received:', message, sender);
   window.postMessage({ type: "BATARANGLE_CONTENT_SCRIPT", message }, "*");
@@ -28,44 +17,6 @@ window.addEventListener("message", function(event) {
     });
   }
 }, false);
-// This script runs before the <head> element is created, so we add the script
-// to <html> instead.
-
-/* Inject our compiled scripts into the inspected application */
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/adapters/angular2.js');
-// document.documentElement.appendChild(script);
-// //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/adapters/base.js');
-// document.documentElement.appendChild(script);
-// // //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/adapters/event_types.js');
-// document.documentElement.appendChild(script);
-// // //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/controllers/base.js');
-// document.documentElement.appendChild(script);
-// // //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/controllers/dom.js');
-// document.documentElement.appendChild(script);
-// // //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/channel/frontend-messaging-service.js');
-// document.documentElement.appendChild(script);
-// // //script.parentNode.removeChild(script);
-
-// var script = document.createElement('script');
-// script.src = chrome.extension.getURL('app/js/batarangle/backend/entry.js');
-// document.documentElement.appendChild(script);
-// //script.parentNode.removeChild(script);
 
 var script = document.createElement('script');
 script.src = chrome.extension.getURL('../../../../node_modules/rx/dist/rx.all.js');
@@ -74,16 +25,8 @@ document.documentElement.appendChild(script);
 
 var url = chrome.extension.getURL('app/js/batarangle/backend/entry.js');
 var imp = 'System.import(\'' + url + '\'); ';
-//console.log(url);
-//console.log(imp);
 
 var script = document.createElement('script');
 script.textContent = imp;
 document.documentElement.appendChild(script);
 //script.parentNode.removeChild(script);
-
-// setTimeout(() => {
-//   window.postMessage({ type: "BATARANGLE_CONTENT_SCRIPT", text: "Start inspecting" }, "*");
-// }, 3000);
-
-
