+++ tsconfig.json
@@ -24,26 +24,20 @@
     "./app/ts/batarangle/frontend/actions/user-actions/user-actions.ts",
 
     "./app/ts/batarangle/frontend/dispatcher/dispatcher.ts",
-
     "./app/ts/batarangle/frontend/stores/abstract-store.ts",
-
     "./app/ts/batarangle/frontend/stores/component-data/component-data-store.ts",
-
     "./app/ts/batarangle/frontend/channel/backend-messaging-service.ts",
 
     "./app/ts/batarangle/channel/channel.ts",
 
-    "./app/ts/batarangle/backend/batarangle-backend.ts",
-    "./app/ts/batarangle/backend/channel/frontend-messaging-service.ts",
-    "./app/ts/batarangle/backend/core/mock-api.ts",
-
-
     "./app/ts/batarangle/backend/adapters/angular2.ts",
     "./app/ts/batarangle/backend/adapters/base.ts",
     "./app/ts/batarangle/backend/adapters/event_types.ts",
-
+    
     "./app/ts/batarangle/backend/controllers/base.ts",
     "./app/ts/batarangle/backend/controllers/dom.ts",
+    
+    "./app/ts/batarangle/backend/batarangle-backend.ts",
     "./app/ts/batarangle/backend/entry.ts"
   ]
 }
